+++ src/components/Layout/Layout.jsx
@@ -11,7 +11,7 @@ import { StaticQuery, graphql } from 'gatsby';
 
 import SEO from '../seo';
 import Header from '../Header/Header';
-import MainMenu from '../menus/mainmenu';
+import MainMenu from '../Menus/MainMenu';
 import Search from '../search/search';
 import Footer from '../Footer/Footer';
 import styles from './layout.module.scss';
