+++ src/components/Menus/mainmenu.module.scss
@@ -1,0 +1,20 @@
+.mainMenu {
+  display: flex;
+  list-style-type: none;
+
+  li {
+    margin-right: 1rem;
+    font-size: 1.8rem;
+    font-weight: bold;
+
+    a,
+    a:link,
+    a:visited {
+      text-decoration: none;
+    }
+  }
+}
+
+.active {
+  color: orange;
+}
