+++ src/components/Menus/socialmenu.module.scss
@@ -1,0 +1,7 @@
+.socialMenu {
+  list-style-type: none;
+
+  li {
+    margin-left: 20%;
+  }
+}
